uniswap-v3-test-tool/metrics/requirements.txt
prometheus-client>=0.16.0
web3>=6.0.0
pyyaml>=6.0
flask>=2.2.0
requests>=2.28.0
python-dotenv>=1.0.0 